services/ssrnet_service/requirements.txt
fastapi>=0.115,<1.0
mtcnn>=1.0,<2.0
numpy>=1.26,<3.0
pillow>=10.4,<12.0
python-multipart>=0.0.9,<1.0
tensorflow==2.19.1
uvicorn>=0.30,<1.0
